web/requirements.txt
alembic
celery==5.0.5
Flask>=1.0.0
Flask-Migrate
Flask-RESTful
Flask-Script
Flask-SQLAlchemy
Jinja2>=2.11.3
SQLAlchemy>=1.3.0
Werkzeug>=0.15.3
pandas
matplotlib
sklearn
pylint-flask
pylint-flask-sqlalchemy
seaborn
flask_login
xgboost
python-dateutil